chore(tests): move unit tests to src directory, update README
Following existing convention in other repos, moving unit tests near the functions they are testing.
