+++ jest.config.js
@@ -2,7 +2,7 @@ module.exports = {
   clearMocks: true,
   globalSetup: './tests/setup.jest.ts',
   globalTeardown: './tests/teardown.jest.ts',
-  setupFilesAfterEnv: ['./tests/unit/mocks.jest.ts'],
+  setupFilesAfterEnv: ['./tests/mocks.jest.ts'],
   moduleFileExtensions: ['ts', 'js', 'mjs'],
   collectCoverageFrom: [
     'src/**/*.ts',
