+++ package.json
@@ -9,8 +9,8 @@
     "format:fix": "prettier . --write",
     "lint:check": "eslint .",
     "lint:fix": "eslint . --fix",
-    "test:unit": "jest --testPathPattern=tests/unit --coverage",
-    "test:integration": "jest --testPathPattern=tests/integration",
+    "test:unit": "jest --testPathPattern=src --coverage",
+    "test:integration": "jest --testPathPattern=tests/**.test.ts --runInBand",
     "test": "yarn build && jest --runInBand",
     "evolve": "yarn ts-node ./tools/evolve-contract.ts",
     "prepare": "husky install",
