+++ tests/balance.test.ts
@@ -1,8 +1,8 @@
 import { Contract, JWKInterface, PstState } from 'warp-contracts';
 
-import { IOState } from '../../src/types';
-import { getLocalArNSContractId, getLocalWallet } from '../utils/helper';
-import { arweave, warp } from '../utils/services';
+import { IOState } from '../src/types';
+import { getLocalArNSContractId, getLocalWallet } from './utils/helper';
+import { arweave, warp } from './utils/services';
 
 describe('Balance', () => {
   let contract: Contract<PstState>;
